Move getBlockByHash to client.ETH
